refactor(parser): break down parser
